Panzer: Fixing exodus reader so that previously recorded fields are preserved
Also modified the stk GatherFields evaluator, it should now be a little more
expressive when it fails. Finally, I had to change the type for the processor
ID to a *double* from an *int*. This was because for reasons I don't understand
that in stk the PROC_ID field previously written by drekar was being shown to
be an double. Not sure why the int type information was getting lost.

+++ packages/panzer/adapters/stk/src/stk_interface/Panzer_STK_Interface.hpp
@@ -91,7 +91,7 @@ buildElementDescriptor(stk::mesh::EntityId elmtId,std::vector<stk::mesh::EntityI
 
 class STK_Interface {
 public:
-   typedef int ProcIdData;
+   typedef double ProcIdData; // ECC: Not sure why?
    typedef stk::mesh::Field<double> SolutionFieldType;
    typedef stk::mesh::Field<double,stk::mesh::Cartesian> VectorFieldType;
    typedef stk::mesh::Field<ProcIdData> ProcIdFieldType;
